Added user_id to post - 12

+++ RestApiCourceTurorial/Controllers/V1/PostsController.cs
@@ -9,6 +9,7 @@ using RestApiCourceTurorial.Contract;
 using RestApiCourceTurorial.Contract.V1.Requests;
 using RestApiCourceTurorial.Contract.V1.Responses;
 using RestApiCourceTurorial.Domain;
+using RestApiCourceTurorial.Extensions;
 using RestApiCourceTurorial.Services;
 
 namespace Tweetbook.Controllers.V1
@@ -31,12 +32,14 @@ namespace Tweetbook.Controllers.V1
         [HttpPut(ApiRoutes.Posts.Update)]
         public async Task<IActionResult> Update([FromRoute] Guid postId, [FromBody] UpdatedPostRequest request)
         {
-            var post = new Post
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+            if (!userOwnsPost)
             {
-                Id = postId,
-                Name = request.Name
+                return BadRequest(new { error = "Youre not own this post" });
+            }
+            var post = await _postService.GetPostByIdAsync(postId);
+            post.Name = request.Name;
 
-            };
             var updated = await _postService.UpdatePostAsync(post);
             if(updated)
                 return Ok(post);
@@ -54,7 +57,10 @@ namespace Tweetbook.Controllers.V1
         [HttpPost(ApiRoutes.Posts.Create)]
         public async Task<IActionResult> Create([FromBody] CreatedPostRequest postRequest)
         {
-            var post = new Post { Name = postRequest.Name };
+            var post = new Post {
+                Name = postRequest.Name,
+                UserId = HttpContext.GetUserId()
+            };
             
             await _postService.CreatePostAsync(post);
 
@@ -68,6 +74,11 @@ namespace Tweetbook.Controllers.V1
         [HttpDelete(ApiRoutes.Posts.Delete)]
         public async Task<IActionResult> Delete([FromRoute] Guid postId)
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+            if (!userOwnsPost)
+            {
+                return BadRequest(new { error = "Youre not own this post" });
+            }
             var deleted = await _postService.DeletePostAsync(postId);
             if (deleted) return NoContent();
             return NotFound();
